brady's suggestions

+++ bindings/cs/libhat-sharp/Native/Signature.cs
@@ -6,5 +6,5 @@ namespace Hat.Native;
 public struct Signature
 {
 	public nint Data;
-	public uint Length;
+	public nint Length;
 }
